requirements.txt
matplotlib
pillow
requests
pyknp
jaconv
